COMMERCE-5392 change package name for system health status classes
