+++ modules/apps/commerce/commerce-health-status-web/src/main/java/com/liferay/commerce/health/status/web/internal/health/status/AccountsCommerceHealthHttpStatus.java
@@ -12,21 +12,28 @@
  * details.
  */
 
-package com.liferay.commerce.health.status.web.internal.util;
+package com.liferay.commerce.health.status.web.internal.health.status;
 
+import com.liferay.commerce.account.constants.CommerceAccountConstants;
+import com.liferay.commerce.account.util.CommerceAccountRoleHelper;
 import com.liferay.commerce.constants.CommerceHealthStatusConstants;
 import com.liferay.commerce.health.status.CommerceHealthHttpStatus;
-import com.liferay.commerce.product.model.CPMeasurementUnit;
-import com.liferay.commerce.product.service.CPMeasurementUnitLocalService;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Role;
+import com.liferay.portal.kernel.service.RoleLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextFactory;
+import com.liferay.portal.kernel.transaction.Propagation;
+import com.liferay.portal.kernel.transaction.TransactionConfig;
+import com.liferay.portal.kernel.transaction.TransactionInvokerUtil;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 
-import java.util.List;
 import java.util.Locale;
 import java.util.ResourceBundle;
+import java.util.concurrent.Callable;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -39,22 +46,31 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	enabled = false, immediate = true,
 	property = {
-		"commerce.health.status.display.order:Integer=100",
-		"commerce.health.status.key=" + CommerceHealthStatusConstants.MEASUREMENT_UNITS_COMMERCE_HEALTH_STATUS_KEY
+		"commerce.health.status.display.order:Integer=110",
+		"commerce.health.status.key=" + CommerceHealthStatusConstants.ACCOUNTS_COMMERCE_HEALTH_STATUS_KEY
 	},
 	service = CommerceHealthHttpStatus.class
 )
-public class MeasurementUnitsCommerceHealthHttpStatus
+public class AccountsCommerceHealthHttpStatus
 	implements CommerceHealthHttpStatus {
 
 	@Override
 	public void fixIssue(HttpServletRequest httpServletRequest)
 		throws PortalException {
 
-		ServiceContext serviceContext = ServiceContextFactory.getInstance(
-			httpServletRequest);
+		try {
+			ServiceContext serviceContext = ServiceContextFactory.getInstance(
+				httpServletRequest);
 
-		_cpMeasurementUnitLocalService.importDefaultValues(serviceContext);
+			Callable<Object> accountRoleCallable = new AccountRoleCallable(
+				serviceContext);
+
+			TransactionInvokerUtil.invoke(
+				_transactionConfig, accountRoleCallable);
+		}
+		catch (Throwable throwable) {
+			_log.error(throwable, throwable);
+		}
 	}
 
 	@Override
@@ -65,13 +81,13 @@ public class MeasurementUnitsCommerceHealthHttpStatus
 		return LanguageUtil.get(
 			resourceBundle,
 			CommerceHealthStatusConstants.
-				MEASUREMENT_UNITS_COMMERCE_HEALTH_STATUS_DESCRIPTION);
+				ACCOUNTS_COMMERCE_HEALTH_STATUS_DESCRIPTION);
 	}
 
 	@Override
 	public String getKey() {
 		return CommerceHealthStatusConstants.
-			MEASUREMENT_UNITS_COMMERCE_HEALTH_STATUS_KEY;
+			ACCOUNTS_COMMERCE_HEALTH_STATUS_KEY;
 	}
 
 	@Override
@@ -81,8 +97,7 @@ public class MeasurementUnitsCommerceHealthHttpStatus
 
 		return LanguageUtil.get(
 			resourceBundle,
-			CommerceHealthStatusConstants.
-				MEASUREMENT_UNITS_COMMERCE_HEALTH_STATUS_KEY);
+			CommerceHealthStatusConstants.ACCOUNTS_COMMERCE_HEALTH_STATUS_KEY);
 	}
 
 	@Override
@@ -95,13 +110,46 @@ public class MeasurementUnitsCommerceHealthHttpStatus
 	public boolean isFixed(long companyId, long commerceChannelId)
 		throws PortalException {
 
-		List<CPMeasurementUnit> cpMeasurementUnits =
-			_cpMeasurementUnitLocalService.getCPMeasurementUnits(companyId);
+		Role role = _roleLocalService.fetchRole(
+			companyId,
+			CommerceAccountConstants.ROLE_NAME_ACCOUNT_ADMINISTRATOR);
 
-		return !cpMeasurementUnits.isEmpty();
+		if (role != null) {
+			return true;
+		}
+
+		return false;
 	}
 
+	private static final Log _log = LogFactoryUtil.getLog(
+		AccountsCommerceHealthHttpStatus.class);
+
+	private static final TransactionConfig _transactionConfig =
+		TransactionConfig.Factory.create(
+			Propagation.REQUIRED, new Class<?>[] {Exception.class});
+
 	@Reference
-	private CPMeasurementUnitLocalService _cpMeasurementUnitLocalService;
+	private CommerceAccountRoleHelper _commerceAccountRoleHelper;
+
+	@Reference
+	private RoleLocalService _roleLocalService;
+
+	private class AccountRoleCallable implements Callable<Object> {
+
+		@Override
+		public Object call() throws Exception {
+			_commerceAccountRoleHelper.checkCommerceAccountRoles(
+				_serviceContext);
+
+			return null;
+		}
+
+		private AccountRoleCallable(ServiceContext serviceContext) {
+			_serviceContext = serviceContext;
+		}
+
+		private final ServiceContext _serviceContext;
+
+	}
 
 }
